add member role. changed all primary key from long to int

+++ src/main/java/com/ralastra/pm_api_spring/controller/UserController.java
@@ -28,7 +28,7 @@ public class UserController {
     }
     
     @GetMapping("/getuserbyid/{id}")
-    public ResponseEntity<UserDTO> getUserById(@PathVariable Long id) {
+    public ResponseEntity<UserDTO> getUserById(@PathVariable Integer id) {
         UserDTO user = userService.getUserById(id);
         return ResponseEntity.ok(user);
     }
@@ -41,14 +41,14 @@ public class UserController {
     
     @PutMapping("/updateuser/{id}")
     public ResponseEntity<UserDTO> updateUser(
-            @PathVariable Long id,
+            @PathVariable Integer id,
             @RequestBody UpdateUserRequest request) {
         UserDTO updatedUser = userService.updateUser(id, request);
         return ResponseEntity.ok(updatedUser);
     }
     
     @DeleteMapping("/deleteuser/{id}")
-    public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteUser(@PathVariable Integer id) {
         userService.deleteUser(id);
         return ResponseEntity.noContent().build();
     }
